move user loader tests
